% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COUNT.R
\name{COUNT}
\alias{COUNT}
\title{Basic COUNT function from excel}
\usage{
COUNT(value)
}
\arguments{
\item{value}{Count amount of the values in the range.}
}
\value{
In this example we count the amount of species in the built-in iris dataset. Function will always return numeric class.
}
\description{
It acts similiarly to Excel's COUNT function. It counts the amount of values in the given array.
}
\examples{
COUNT(iris$Species)
}
