% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COUNTIF.R
\name{COUNTIF}
\alias{COUNTIF}
\title{Basic COUNTIF function from excel}
\usage{
COUNTIF(range, criteria)
}
\arguments{
\item{range}{Specify range for Countif}

\item{criteria}{Give the criteria to check the range for.}
}
\value{
In this example we count the amount of setosa in iris dataset. Function will always return numeric class.
}
\description{
It acts similiarly to Excel's COUNTIF function. It counts the amount of cells that comply with the given criteria.
}
\examples{
COUNTIF(iris$Species,"setosa")
}
