% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COUNTIFS.R
\name{COUNTIFS}
\alias{COUNTIFS}
\title{Basic COUNTIFS function from excel}
\usage{
COUNTIFS(
  criteria_range1,
  criteria1,
  criteria_range2 = TRUE,
  criteria2 = TRUE,
  criteria_range3 = TRUE,
  criteria3 = TRUE,
  criteria_range4 = TRUE,
  criteria4 = TRUE,
  criteria_range5 = TRUE,
  criteria5 = TRUE,
  criteria_range6 = TRUE,
  criteria6 = TRUE,
  criteria_range7 = TRUE,
  criteria7 = TRUE,
  criteria_range8 = TRUE,
  criteria8 = TRUE,
  criteria_range9 = TRUE,
  criteria9 = TRUE,
  criteria_range10 = TRUE,
  criteria10 = TRUE
)
}
\arguments{
\item{criteria_range1, criteria_range2, criteria_range3, criteria_range4, criteria_range5, criteria_range6, criteria_range7, criteria_range8, criteria_range9, criteria_range10}{Specify range for Countifs, only criteria_range1 is mandatory.}

\item{criteria1, criteria2, criteria3, criteria4, criteria5, criteria6, criteria7, criteria8, criteria9, criteria10}{Give the criteria to check the range for.Only criteria1 is necessary, others are optional.}
}
\value{
In this example we count the amount of cells where Spcies are setosa and has a Petal Width of 0.2 .Function will always return numeric class.
}
\description{
It acts similiarly to Excel's COUNTIFS function. Counts values in a range which comply with given criteria.
}
\examples{
COUNTIFS(iris$Species,"setosa",iris$Petal.Width,0.2)
}
