% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATE.R
\name{DATE}
\alias{DATE}
\title{Basic DATE function from excel}
\usage{
DATE(year, month, day)
}
\arguments{
\item{year}{Give year argument to the function.}

\item{month}{Give month argument to the function.}

\item{day}{Give day argument to the function.}
}
\value{
This example returns 23rd June of 2020. Function will always return Date class.
}
\description{
It acts similiarly to Excel's DATE function. You give 3 arguments which are year,month and day and it will give you the date in a date format.
}
\examples{
DATE(2020,23,06)
}
