% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATEDIF.R
\name{DATEDIF}
\alias{DATEDIF}
\title{Basic DATEDIF function from excel}
\usage{
DATEDIF(start_date, end_date, difference = "d")
}
\arguments{
\item{start_date}{Start date to evaluate the difference}

\item{end_date}{End Date to evaluate the difference}

\item{difference}{What type of difference do you want? Year,Month or Day? Specify "m" for example for month/months, "d" for day/days and "y" for year/years.}
}
\value{
In these examples we have all 3 types of returns, first is difference between specified two dates in days, second one is difference in months and third one is difference in years.Function will always return numeric class.
}
\description{
It acts similiarly to Excel's DATEDIF function. It returns difference between two dates, either day,month or year, it's up to the user to specify which type of difference user wants.
}
\examples{
DATEDIF(DATE(2020,1,1),DATE(2020,2,1),"d")
DATEDIF(DATE(2020,1,1),DATE(2020,2,1),"m")
DATEDIF(DATE(2020,1,1),DATE(2020,2,1),"Y")
}
