% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATETOEXCELSERIES.R
\name{DATETOEXCELSERIES}
\alias{DATETOEXCELSERIES}
\title{Date to excel date series function}
\usage{
DATETOEXCELSERIES(date)
}
\arguments{
\item{date}{Convert R date type to Excel general date series, this might be helpful for Excel users.}
}
\value{
In this example it returns the Excel's general date series equivalent of date 1st January of 2020. Function will always return numeric class.
}
\description{
Functions converts dates to Excel General date series which might be useful when writing Excel files.
}
\examples{
DATETOEXCELSERIES(DATE(2020,1,1))
}
