% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAY.R
\name{DAY}
\alias{DAY}
\title{Basic DAY function from excel}
\usage{
DAY(date)
}
\arguments{
\item{date}{Give the date argument so it can extract day from the date.}
}
\value{
in this example we have 13th of January. Function will return 13 as it is the day from the date. Function will always return numeric class.
}
\description{
It acts similiarly to Excel's DAY function. It gives you the day from specified date.
}
\examples{
DAY(DATE(2020,1,13))
}
