% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAYS.R
\name{DAYS}
\alias{DAYS}
\title{Basic DAYS function from excel}
\usage{
DAYS(start_date, end_date)
}
\arguments{
\item{start_date}{Give the start_date argument so it can calculate days.}

\item{end_date}{Give the end_date argument so it can calculate days.}
}
\value{
In this example we are interested how many days there are between 1st February 2020 and 15th February 2020 which is 14. Function will always return numeric class.
}
\description{
It acts similiarly to Excel's DAYS function. It calculates the difference between two dates in days.
}
\examples{
DAYS(DATE(2020,2,1),DATE(2020,2,15))
}
