% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EOMONTH.R
\name{EOMONTH}
\alias{EOMONTH}
\title{Basic EOMONTH function from excel}
\usage{
EOMONTH(date, months = 0)
}
\arguments{
\item{date}{Give the date argument so it can give you the end of the month.}

\item{months}{The number of months before or after start_date. A positive value for months yields a future date; a negative value yields a past date.}
}
\value{
In this case we specify 2nd June 2008. Function returns end of the month which is 30th June 2008. Function will always return Date class.
}
\description{
It acts similiarly to Excel's EOMONTH function. It returns the end of month date for the specified date.
}
\examples{
EOMONTH(DATE(2008,6,2))
}
