% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIND.R
\name{FIND}
\alias{FIND}
\title{Basic FIND function from excel}
\usage{
FIND(find_text, within_text)
}
\arguments{
\item{find_text}{Find the text in the text.}

\item{within_text}{Where should the function find the text.}
}
\value{
in this example we try to find on which place does CRAN start. Function will always return numeric class.
}
\description{
It acts similiarly to Excel's FIND function. It finds the starting point of the string where it matches your find_text value.
}
\examples{
FIND("CRAN","I LOVE CRAN")
}
