% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INDEX.R
\name{INDEX}
\alias{INDEX}
\title{Basic INDEX function from excel}
\usage{
INDEX(array, row_num, column_num = 1)
}
\arguments{
\item{array}{Which array/table should it use?}

\item{row_num}{Which row should it return the value from?}

\item{column_num}{Which column should it return the value from?}
}
\value{
In this example we get 3rd row and 2nd column from the dataframe. This function can return different classes numeric, character, logical etc. It depends on what is in array/dataframe.
}
\description{
It acts similiarly to Excel's INDEX function. It gives you the value from dataframe when you specify the array indices(row and column)
}
\examples{
INDEX(iris,3,2)
}
