% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISEVEN.R
\name{ISEVEN}
\alias{ISEVEN}
\title{Basic ISEVEN function from excel}
\usage{
ISEVEN(number)
}
\arguments{
\item{number}{Input the number for it to evaluate if it is even?}
}
\value{
First example returns true as it is an even number 2, second example returns false as it isn't an even number. Function will always return logical class.
}
\description{
It acts similiarly to Excel's ISEVEN function. If the specified number is even then it returns true, if not then false.
}
\examples{
ISEVEN(2)
ISEVEN(1)
}
