% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISLOGICAL.R
\name{ISLOGICAL}
\alias{ISLOGICAL}
\title{Basic ISLOGICAL function from excel}
\usage{
ISLOGICAL(value)
}
\arguments{
\item{value}{Input the number for it to evaluate if it is logical? Works on vectors/arrays as well.}
}
\value{
We have 3 cases in the examples. First one is logical therefore function returns true, second one is also logical and it returns true as well. Third example isn't logical therefore function returns false. Function will always return logical class.
}
\description{
It acts similiarly to Excel's ISLOGICAL function. If specified value is true or false then it returns true, if not then it returns false.
}
\examples{
ISLOGICAL(TRUE)
ISLOGICAL(FALSE)
ISLOGICAL("Is this a logical?")
}
