% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISNUMBER.R
\name{ISNUMBER}
\alias{ISNUMBER}
\title{Basic ISNUMBER function from excel}
\usage{
ISNUMBER(value)
}
\arguments{
\item{value}{Input the number for it to evaluate if it is number? Works on vectors/arrays as well.}
}
\value{
first example returns true as it is a number. Second example returns false as it isn't a number, it's a string. Function will always return logical class.
}
\description{
It acts similiarly to Excel's ISNUMBER function. If the specified value is a number it returns true, in all other cases it returns false.
}
\examples{
ISNUMBER(2)
ISNUMBER("2")
}
