% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISODD.R
\name{ISODD}
\alias{ISODD}
\title{Basic ISODD function from excel}
\usage{
ISODD(number)
}
\arguments{
\item{number}{Input the number for it to evaluate if it is an odd number? Works on vectors/arrays as well.}
}
\value{
First example returns true as it is an odd number 1, second example returns false as it isn't an odd number. Function will always return logical class.
}
\description{
It acts similiarly to Excel's ISODD function. If the specified number is odd then it returns true, if not then false.
}
\examples{
ISODD(1)
ISODD(2)
}
