% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEFT.R
\name{LEFT}
\alias{LEFT}
\title{Basic LEFT function from excel}
\usage{
LEFT(text, num_chars)
}
\arguments{
\item{text}{the text you want to select characters from left.}

\item{num_chars}{How many characters should it select?}
}
\value{
In this case we have a sentence and we want to extract first 4 characters from the sentence. Therefore we specify the argument 4 and it gives us the first word.Function will always return character class.
}
\description{
It acts similiarly to Excel's LEFT function. It takes the text and gives you the amount of characters you want to get from the string.
}
\examples{
LEFT("Fear what happens",4)
}
