% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEN.R
\name{LEN}
\alias{LEN}
\title{Basic LEN function from excel}
\usage{
LEN(text)
}
\arguments{
\item{text}{amount of characters in the word.}
}
\value{
in this example we see how long the sentence is. Function will always return numeric class.
}
\description{
It acts similiarly to Excel's LEN function. This function gives you the length of a string.
}
\examples{
LEN("This is great!")
}
