% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOWER.R
\name{LOWER}
\alias{LOWER}
\title{Basic LOWER function from excel}
\usage{
LOWER(text)
}
\arguments{
\item{text}{Give the function a word to make it lower.Give the texts via vector if you want to perform it on multiple texts.}
}
\value{
In this case we lower the whole specified sentence and return the sentence in all lower characters. Function will always return character class.
}
\description{
It acts similiarly to Excel's LOWER function. It converts the sentence/word to lowercase characters.
}
\examples{
LOWER("THIS IS SPARTAA! IS IT THOUGH AFTER LOWERING?")
}
