% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATCH.R
\name{MATCH}
\alias{MATCH}
\title{Basic MATCH function from excel}
\usage{
MATCH(lookup_value, lookup_array)
}
\arguments{
\item{lookup_value}{what value to lookup}

\item{lookup_array}{Where should it lookup the value}
}
\value{
This example gives us the first index of an array where Species is virginica. Function will always return numeric.
}
\description{
It acts similiarly to Excel's MATCH function. It matches the value in the array.
}
\examples{
MATCH("virginica",iris$Species)
}
