% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAXIF.R
\name{MAXIF}
\alias{MAXIF}
\title{Basic MAXIF function from excel}
\usage{
MAXIF(range, criteria, max_range)
}
\arguments{
\item{range}{Range where it should check the criteria}

\item{criteria}{Where should it lookup the value}

\item{max_range}{Which array should it return the max from.}
}
\value{
In this case we get the maximum value of Sepal Length from Species which are virginica. Function will always return numeric class.
}
\description{
It acts similiarly to Excel's MAXIF function. It returns the maximum value from an array after testing for certain criterias.
}
\examples{
MAXIF(iris$Species,"virginica",iris$Sepal.Length)
}
