% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MID.R
\name{MID}
\alias{MID}
\title{Basic MID function from excel}
\usage{
MID(text, start_num, num_chars)
}
\arguments{
\item{text}{From whch text should it return the string?}

\item{start_num}{Where should it start counting from?}

\item{num_chars}{How many characters should it return?}
}
\value{
In this example we want to extract "kata" from this string. So we specify 5 as the starting number and 4 as the amount of characters to extract. Function will always return character class.
}
\description{
It acts similiarly to Excel's MID function. Function is for string extraction. You select the starting number and the amount of characters you want to extract.
}
\examples{
MID("Kayakata",5,4)
}
