% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINIF.R
\name{MINIF}
\alias{MINIF}
\title{Basic MINIF function from excel}
\usage{
MINIF(range, criteria, min_range)
}
\arguments{
\item{range}{Which range should it check the criteria against?}

\item{criteria}{What should be checked?}

\item{min_range}{From which range should it return the minimum from?}
}
\value{
In this case we get the minimum value of Sepal Length from Species which are virginica. Function will always return numeric class.
}
\description{
It acts similiarly to Excel's MINIF function. It returns the minimum value from an array after testing for certain criterias.
}
\examples{
MINIF(iris$Species,"virginica",iris$Sepal.Length)
}
