% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MINIFS.R
\name{MINIFS}
\alias{MINIFS}
\title{Basic MINIFS function from excel}
\usage{
MINIFS(
  min_range,
  criteria_range1,
  criteria1,
  criteria_range2 = TRUE,
  criteria2 = TRUE,
  criteria_range3 = TRUE,
  criteria3 = TRUE,
  criteria_range4 = TRUE,
  criteria4 = TRUE,
  criteria_range5 = TRUE,
  criteria5 = TRUE,
  criteria_range6 = TRUE,
  criteria6 = TRUE,
  criteria_range7 = TRUE,
  criteria7 = TRUE,
  criteria_range8 = TRUE,
  criteria8 = TRUE,
  criteria_range9 = TRUE,
  criteria9 = TRUE,
  criteria_range10 = TRUE,
  criteria10 = TRUE
)
}
\arguments{
\item{min_range}{From which range should it return the minimum from?}

\item{criteria_range1, criteria_range2, criteria_range3, criteria_range4, criteria_range5, criteria_range6, criteria_range7, criteria_range8, criteria_range9, criteria_range10}{Which range should the criteria tested for.Only criteria_range1 is mandatory,others are optional.}

\item{criteria1, criteria2, criteria3, criteria4, criteria5, criteria6, criteria7, criteria8, criteria9, criteria10}{What criteria should the range be checked against. Only criteria1 is mandatory, others are optional.}
}
\value{
This example returns minimum value of Petal Length of species setosa who have petal width 0.2. Function will always return numeric class.
}
\description{
It acts similiarly to Excel's MINIFS function. It returns the minimum value from an array after testing for several criterias.
}
\examples{
MINIFS(iris$Petal.Length,iris$Species,"setosa",iris$Petal.Width,0.2)
}
