% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOT.R
\name{NOT}
\alias{NOT}
\title{Basic NOT function from excel}
\usage{
NOT(logical)
}
\arguments{
\item{logical}{Enter the logical to get the opposite logical of it. For example if you input TRUE, it will get FALSE.}
}
\value{
in the first example it will return false while in the 2nd example it will return true. Function will always return logical class.
}
\description{
It acts similiarly to Excel's NOT function. It returns the opposite of the logical you specify.
}
\examples{
NOT(TRUE)
NOT(FALSE)
}
