% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOW.R
\name{NOW}
\alias{NOW}
\title{Basic NOW function from excel}
\usage{
NOW()
}
\value{
As the function has no arguments it simply returns current system time in character format. Function will always return character class.
}
\description{
It acts similiarly to Excel's NOW function. It gives the system time in character format.
}
\examples{
NOW()
}
