% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAND.R
\name{RAND}
\alias{RAND}
\title{Basic RAND function from excel.}
\usage{
RAND()
}
\value{
This example simply returns a number from 0 to 1. Function will always return numeric class.
}
\description{
It acts similiarly to Excel's RAND function. No need to specify the arguments/parameters. It gives you the random number from 0 to 1.
}
\examples{
RAND()
}
