% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RANDBETWEEN.R
\name{RANDBETWEEN}
\alias{RANDBETWEEN}
\title{Basic RANDBETWEEN function from excel}
\usage{
RANDBETWEEN(bottom, top, number = 1)
}
\arguments{
\item{bottom}{Give the function a bottom floor for the randbetween}

\item{top}{Give the function a top ceiling for the randbetween}

\item{number}{How many numbers should it generate?}
}
\value{
In the first example we get only 1 number from 1 to 100, while in the second example we get 3 numbers from 1 to 100 as the argument number is specified 3. Function will always return numeric class.
}
\description{
It acts similiarly to Excel's RANDBETWEEN function. it takes several arguments like bottom,top and number, you specify the floor, ceiling and the amount of numbers you want to generate and it gives you the random between the floor and ceiling.
}
\examples{
RANDBETWEEN(1,100, number = 1)
RANDBETWEEN(1,100, number = 3)
}
