% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REPT.R
\name{REPT}
\alias{REPT}
\title{Basic REPT function from excel}
\usage{
REPT(text, number_times, AsOne = TRUE)
}
\arguments{
\item{text}{Which text should it repeat n time?}

\item{number_times}{How many times should the function repeat the given text.}

\item{AsOne}{Should function concatenate the text or should it return seperately as a vector(Vector is same as array in Excel)}
}
\value{
In the first example we repeat "Oi" 2 times and it is coerced together as one string. In the second example we don't coerce it together but it still repeats the "Oi" two times. Function will always return character class.
}
\description{
It acts similiarly to Excel's REPT function. Repeat the text as many times as you want.
}
\examples{
REPT("Oi",2,AsOne = TRUE)
REPT("Oi",2,AsOne = FALSE)
}
