% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIGHT.R
\name{RIGHT}
\alias{RIGHT}
\title{Basic RIGHT function from excel}
\usage{
RIGHT(text, num_chars)
}
\arguments{
\item{text}{from where should it get the characters}

\item{num_chars}{how many characters should it get?}
}
\value{
In this example we take "Kayakata" and extract 4 characters from the right handside of the string. Functions will always return character class.
}
\description{
It acts similiarly to Excel's RIGHT function. It takes the string and takes the amount of characters you want to extract from it.
}
\examples{
RIGHT("Kayakata",4)
}
