% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sales.R
\docType{data}
\name{Sales}
\alias{Sales}
\title{Random Sales Data}
\format{
A data frame of 24 rows and 4 columns
\describe{
 \item{Names}{Names of salesman}
 \item{Country}{Countries of the salesman}
 \item{Cost}{Cost of each salesman}
 \item{Sales}{Amount of sales each salesman generates}
}
}
\source{
Randomly generated data
}
\usage{
Sales
}
\description{
A dataset containing randomly generated Sales data.
}
\keyword{datasets}
