% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Streets.R
\docType{data}
\name{Streets}
\alias{Streets}
\title{Random Salesman Streets Data}
\format{
A data frame of 4 rows and 2 columns
\describe{
 \item{Country}{Names of salesman}
 \item{Street}{Street the salesman lives on}
}
}
\source{
Randomly generated data
}
\usage{
Streets
}
\description{
A dataset containing randomly generated Streets data.
}
\keyword{datasets}
