% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUBSTITUTE.R
\name{SUBSTITUTE}
\alias{SUBSTITUTE}
\title{Basic SUBSTITUTE function from excel}
\usage{
SUBSTITUTE(text, old_text, new_text)
}
\arguments{
\item{text}{Where should it substitute the characters}

\item{old_text}{Which text should it substitute}

\item{new_text}{What should it substitute with.}
}
\value{
In this example we take text "CRAN", we take the old text "RAN" and replace it with "out" which in return gives us "Cout". Function will always return character class.
}
\description{
It acts similiarly to Excel's SUBSTITUTE function. If you want to substitute the characters by certain characters you should use this function.
}
\examples{
SUBSTITUTE("CRAN","RAN","out")
}
