% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUMIF.R
\name{SUMIF}
\alias{SUMIF}
\title{Basic SUMIF function from excel}
\usage{
SUMIF(range, criteria, sum_range)
}
\arguments{
\item{range}{Which range should it check the criteria against.}

\item{criteria}{what criteria should it check in range}

\item{sum_range}{Which range should it sum}
}
\value{
In this case we are summing Sepal length of species which are virginica. Function will always return numeric class.
}
\description{
It acts similiarly to Excel's SUMIF function. It sums the values where certain criterias are met.
}
\examples{
SUMIF(iris$Species,"virginica",iris$Sepal.Length)
}
