% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TODAY.R
\name{TODAY}
\alias{TODAY}
\title{Basic TODAY function from excel}
\usage{
TODAY()
}
\value{
Example returns the system date as the function does, nothing specific. Function will always return Date class.
}
\description{
It acts similiarly to Excel's TODAY function.No need to give the arguments. Function returns the system date.
}
\examples{
TODAY()
}
