% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UPPER.R
\name{UPPER}
\alias{UPPER}
\title{Basic UPPER function from excel}
\usage{
UPPER(text)
}
\arguments{
\item{text}{Give this function the text to capitalize all the letters.Give this function words with a vector if you want to perform it on several texts.}
}
\value{
In this case we have specified "is this sparta?" and it has returned all the characters in uppercase as expected. Function will always return character class.
}
\description{
It acts similiarly to Excel's UPPER function. It takes the string and coverts all of it's characters to uppercase.
}
\examples{
UPPER("is this sparta?")
}
