% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VLOOKUP.R
\name{VLOOKUP}
\alias{VLOOKUP}
\title{Basic VLOOKUP function from excel}
\usage{
VLOOKUP(
  lookup_from_table,
  lookup_column_value = "Name of the column to lookup",
  lookup_where_table,
  lookup_where_table_column = "Name of the column to compare",
  return_which_column = "Name of the column to return"
)
}
\arguments{
\item{lookup_from_table}{The table it should lookup values from}

\item{lookup_column_value}{which column should be looked up}

\item{lookup_where_table}{which table should it look for the values in}

\item{lookup_where_table_column}{Which column should it look for the values in.}

\item{return_which_column}{Which column should it return}
}
\value{
In this case we have built-in database Sales and Street. We try to merge these 2 tables to see on which street are the salesman based on their countries. Function can return numeric,character,logical or any other class, it depends on what is in the table you are looking up the value in.
}
\description{
It acts similiarly to Excel's VLOOKUP function with some extra arguments. It takes the value that you want to take from another table and returns the corresponding value from another table. Basically it's an SQL Left Join.
}
\examples{
VLOOKUP(Sales,"Country",Streets,"Country","Street")
}
