% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YEAR.R
\name{YEAR}
\alias{YEAR}
\title{Basic YEAR function from excel}
\usage{
YEAR(date)
}
\arguments{
\item{date}{Give the date argument so it can extract year from the date. Preferable to give the date via DATE function of this package.}
}
\value{
In this example function returns 2020 as it is the year part of the date specified. Function will always return numeric class..
}
\description{
It acts similiarly to Excel's YEAR function. Function will extract year component of your date.
}
\examples{
YEAR(DATE(2020,1,1))
}
